requirements.txt
jinja2==2.11.2
pyyaml==5.3.1
pytest==6.1.2